refactor: moves blog creator plugin to Layout
