+++ pages/blog/[slug].tsx
@@ -1,8 +1,6 @@
 import styled from 'styled-components'
 const fg = require('fast-glob')
 import { NextSeo } from 'next-seo'
-import { usePlugin } from 'tinacms'
-import { MarkdownCreatorPlugin } from '../../utils/plugins'
 
 import { formatDate, formatExcerpt } from '../../utils'
 import {
@@ -42,59 +40,16 @@ export default function BlogTemplate({
     editMode
   )
 
-  const CreateBlogPlugin = new MarkdownCreatorPlugin({
-    label: 'New Blog Post',
-    filename: form => {
-      const slug = form.title.replace(/\s+/, '-').toLowerCase()
-      return `content/blog/${slug}.md`
-    },
-    fields: [
-      {
-        name: 'title',
-        component: 'text',
-        label: 'Title',
-        placeholder: 'My New Post',
-        description: 'The title of the new blog post.',
-      },
-      {
-        label: 'Date',
-        name: 'date',
-        component: 'date',
-        description: 'The default will be today',
-      },
-      {
-        label: 'Author',
-        description: 'Who wrote this, yo?',
-        name: 'author',
-        component: 'text',
-      },
-    ],
-    githubOptions: sourceProviderConnection,
-    isEditMode: editMode,
-    frontmatter: postInfo => ({
-      title: postInfo.title,
-      date: postInfo.date ? postInfo.date : new Date(),
-      author: postInfo.author ? postInfo.author : `Jane Doe`,
-    }),
-    body: postInfo => `New post, who dis?`,
-    afterCreate: response => {
-      let url = fileToUrl(
-        response.data.content.path.split('content')[1],
-        'blog'
-      )
-      window.location.href = url
-    },
-  })
-
-  usePlugin(CreateBlogPlugin)
-
   const frontmatter = data.frontmatter
   const markdownBody = data.markdownBody
   const excerpt = formatExcerpt(data.markdownBody)
 
   return (
     <OpenAuthoringSiteForm form={form} editMode={editMode}>
-      <Layout pathname="/">
+      <Layout
+        sourceProviderConnection={sourceProviderConnection}
+        editMode={editMode}
+      >
         <NextSeo
           title={frontmatter.title}
           titleTemplate={'%s | ' + siteConfig.title + ' Blog'}
