+++ pages/teams.tsx
@@ -58,7 +58,11 @@ export default function TeamsPage(props) {
 
   return (
     <OpenAuthoringSiteForm form={form} editMode={props.editMode}>
-      <TeamsLayout page="teams" color={'secondary'}>
+      <TeamsLayout
+        sourceProviderConnection={props.sourceProviderConnection}
+        editMode={props.editMode}
+        color={'secondary'}
+      >
         <NextSeo
           title={data.title}
           description={data.description}
